WSC203/requirements.txt
-i https://pypi.org/simple
babel==2.6.0
click==6.7
decorator==4.3.0
dominate==2.3.1
flask-babel==0.11.2
flask-bootstrap4==4.0.2
flask-login==0.4.1
flask-moment==0.6.0
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==1.0.2
gpio==0.2.0
itsdangerous==0.24
jinja2==2.10
markupsafe==1.1.1
netifaces==0.10.7
paho-mqtt==1.3.1
pbr==4.2.0
pycryptodomex==3.6.3
pyserial==3.4
pytz==2018.5
six==1.11.0
sqlalchemy-migrate==0.11.0
sqlalchemy==1.2.11
sqlparse==0.2.4
tempita==0.5.2
uwsgi==2.0.17.1
visitor==0.1.3
werkzeug==0.16.1
wtforms==2.2.1
